Refactor <FilterPopupButton /> component out of the <TransactionFilter />
