+++ frontend/src/Views/TransactionList/TransactionFilter.tsx
@@ -1,7 +1,7 @@
-import { Button, Fab, Grid, Popover, Typography } from "@material-ui/core";
-import { FilterList } from "@material-ui/icons";
+import { Button, Grid, Typography } from "@material-ui/core";
 import { makeStyles } from "@material-ui/styles";
 import React, { useState } from "react";
+import FilterPopupButton from "../FilterPopupButton";
 import FilterSelect from "../FilterSelect";
 
 const useStyles = makeStyles(() => ({
@@ -26,10 +26,7 @@ const tags = ["foo", "bar", "baz", "bez", "booze", "bamboozle"];
 const TransactionFilter = () => {
   const classes = useStyles();
 
-  const [anchorEl, setAnchorEl] = useState<any>(null);
   const [filters, setFilters] = useState<IFilters>(emptyFilterObject);
-
-  const open = !!anchorEl;
 
   // TODO: The type annotations for the event parameter dont work
   // Typescript thinks that event.target.value is a `string` while
@@ -42,58 +39,32 @@ const TransactionFilter = () => {
     }));
   };
 
-  const handleClick = (event: React.MouseEvent<HTMLElement, MouseEvent>) => {
-    setAnchorEl(event.target);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
   return (
-    <React.Fragment>
-      <Fab onClick={handleClick} color="secondary" aria-label="Filter">
-        <FilterList />
-      </Fab>
-      <Popover
-        id="filter-popper"
-        open={open}
-        anchorEl={anchorEl}
-        onClose={handleClose}
-        anchorOrigin={{
-          vertical: "center",
-          horizontal: "center"
-        }}
-        transformOrigin={{
-          vertical: "bottom",
-          horizontal: "right"
-        }}
-      >
-        <div className={classes.filterPopup}>
-          <Grid container={true} spacing={16}>
-            <Grid item={true} xs={3}>
-              <Typography variant="h5">Filters</Typography>
-            </Grid>
-            <Grid item={true} xs={9}>
-              <Button color="primary">
-                <Typography color="primary" variant="button">
-                  Reset
-                </Typography>
-              </Button>
-            </Grid>
-
-            <Grid item={true} xs={12}>
-              <FilterSelect
-                label="Tags"
-                selected={filters.tags}
-                items={tags}
-                handleChange={handleChange("tags")}
-              />
-            </Grid>
+    <FilterPopupButton>
+      <div className={classes.filterPopup}>
+        <Grid container={true} spacing={16}>
+          <Grid item={true} xs={3}>
+            <Typography variant="h5">Filters</Typography>
           </Grid>
-        </div>
-      </Popover>
-    </React.Fragment>
+          <Grid item={true} xs={9}>
+            <Button color="primary">
+              <Typography color="primary" variant="button">
+                Reset
+              </Typography>
+            </Button>
+          </Grid>
+
+          <Grid item={true} xs={12}>
+            <FilterSelect
+              label="Tags"
+              selected={filters.tags}
+              items={tags}
+              handleChange={handleChange("tags")}
+            />
+          </Grid>
+        </Grid>
+      </div>
+    </FilterPopupButton>
   );
 };
